// server/routes/modules/assistantModule.types.ts
export type AssistantModuleDeps = {
  addCallsToThreadInMicroBatches: any;
  checkAdminAccess: any;
  db: any;
  eq: any;
  getEffectiveTenantId: any;
  googleDrive: any;
  isAdmin: any;
  isAuthenticated: any;
  isAuthenticatedCustom: any;
  kbFiles: any;
  sql: any;
  storage: any;
  syncKbFileToAlignerVectorStore: any;
};